% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.probe.vec.R
\name{annot.probe.vec}
\alias{annot.probe.vec}
\title{Title Annotate vector of Illumina probe names}
\usage{
annot.probe.vec(probe.vec, annot = NULL, annotation.file.name = NULL,
  required.annotation = c("IlmnID", "Coordinate_37", "UCSC_RefGene_Name",
  "UCSC_RefGene_Group", "UCSC_CpG_Islands_Name",
  "Relation_to_UCSC_CpG_Island"))
}
\arguments{
\item{probe.vec}{A vector of Illumina probe names}

\item{annot}{A preloaded data.table of annotation}

\item{annotation.file.name}{If annot is not given, annotation.file.name can provide the file name of the illumine annotation file and the function will load and convert it to a data.table. 
If neither annot nor annotation.file.name are provided, annotation will be loaded from Tim Triche's bioconductor R package.}

\item{required.annotation}{Names of columns from Illumina annotation file, to be provided as annotation}
}
\value{
A matrix of requred.annotations for each of the probes in probe.vec
}
\description{
Returns a matrix with annotations of the CpG sites represented by a given vector of probe names
}
\examples{

data(betas.7)
data(annot.7)
annot.sites <- annot.probe.vec(rownames(betas.7)[1:10], annot = annot.7)


}
