% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.annot.triche.R
\name{create.annot.triche}
\alias{create.annot.triche}
\title{Title Create an annotation to a vector of probes}
\usage{
create.annot.triche(probe.vec, only.locations = F)
}
\arguments{
\item{probe.vec}{A vector of Illumina probe names}

\item{only.locations}{If TRUE, only chromosomes and CpG location will be returned.}
}
\value{
A data table with annotation information. If only.locations=T, chromosome and chromosomal locations are given. Otherwise, also infinium design type, gene name, functional group, associated island (or CpG resort), type of region in the resort.
}
\description{
Title Create an annotation to a vector of probes
}
\examples{


data(betas.7)
source("http://bioconductor.org/biocLite.R")  
biocLite("IlluminaHumanMethylation450k.db")
require(IlluminaHumanMethylation450k.db) # this must be loaded
annot.7 <- create.annot.triche(rownames(betas.7), only.locations = T)

}
