% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dbp.merge.R
\name{Dbp.merge}
\alias{Dbp.merge}
\title{Title Dbp.merge}
\usage{
Dbp.merge(ordr.vec, thresh.dist, bp.thresh.dist, location.vec,
  dist.type = "spearman")
}
\arguments{
\item{ordr.vec}{An (n by m) matrix of n measurements of m random variables (methylations). The columns are ordered.}

\item{thresh.dist}{A similarity distance threshold. Sites in the interval are merged to a single cluster if the similarity distance between them is above thresh.dist.}

\item{bp.thresh.dist}{A threshold distance in (e.g.) chromosomal location. Any interval to be potentially merged is smaller or equal to bp.thresh.dist.}

\item{location.vec}{An m vector specifying the locations (e.g. chromosomal locations) of the variables measured in the matrix ordr.vec.}

\item{dist.type}{Type of similarity distance function "spearman", "person" or "euclid".}
}
\value{
An m vector of cluster assignments of the m ordered variables.
}
\description{
Merge all sites on an interval, if the sites at the ends of the interval are ``close" to each other
}
\examples{



}
