% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{first}
\alias{first}
\title{Title The first entry in a vector}
\usage{
first(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The value in the first entry in x (i.e x[1])
}
\description{
Title The first entry in a vector
}
