% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Title The last entry in a vector}
\usage{
last(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The value in the last entry in x
}
\description{
Returns the value in the last entry in a vector
}
