% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.betas.by.chrom.location.R
\name{order.betas.by.chrom.location}
\alias{order.betas.by.chrom.location}
\title{Title Organizes betas by location}
\usage{
order.betas.by.chrom.location(betas, annot = NULL,
  annotation.file.name = NULL, return.chroms = NULL)
}
\arguments{
\item{betas}{An (m by n) matrix of methylation values measured on $n$ participants in $m$ sites.}

\item{annot}{A preloaded data.table of annotation}

\item{annotation.file.name}{If annot is not given, annotation.file.name can provide the file name of the illumine annotation file and the function will load and convert it to a data.table. 
If neither annot nor annotation.file.name are provided, annotation will be loaded from Tim Triche's bioconductor R package.}

\item{return.chroms}{Optional list of chromosomes, if one is interested in specific chromosomes.}
}
\value{
\item{betas.by.chrom}{A list ordered by chromosome number. Each item in this list contains a matrix of methylation values for the subsets of sites from betas from the corresponding chromosome. The rows are organized according to chromosomal location. }
\item{sites.locations.by.chrom}{A list ordered by chromosome number. Each item in this list contains a matrix specifying chromosomal locations of each of the sites in the subsets of sites from betas from the corresponding chromosome. The rows are organized according to chromosomal location. }
}
\description{
Organizes a matrix of methylation values to be ordered by chromosomal location
}
\examples{

data(betas.7) ## upload methylation data
data(annot.7)
dat.7.ord <- order.betas.by.chrom.location(betas.7, annot = annot.7)
dat.7.ord$betas.by.chrom$"7"[1:5,1:5]
dat.7.ord$sites.locations.by.chrom$"7"[1:5]

}
