% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cor.clusts.R
\name{plot.cor.clusts}
\alias{plot.cor.clusts}
\title{Title Plots clusters on a heatmap of a correlation matrix}
\usage{
plot.cor.clusts(cor.mat, clusts, erase.non.clust = F, lwd = 2,
  main = "", xlab = "", labels = NULL)
}
\arguments{
\item{cor.mat}{An (m by m) correlation matrix}

\item{clusts}{An $m$ vector of clusters assignments, as output by Acluster function.}

\item{erase.non.clust}{If TRUE, do not present correlations between sites that are not in the same cluster (i.e. they are just white). Default is FALSE.}

\item{lwd}{The width of the lines marking the clusters}

\item{main}{Main title of the plot}

\item{xlab}{Label of the X-axis}

\item{labels}{Labels of the sites}
}
\value{
A heatmap
}
\description{
Given a correlation matrix and a vector of cluster assignments, plots squares around clusters of at least 2 sites.
}
\examples{

data(betas.7)
data(annot.7)
dat.7.ord <- order.betas.by.chrom.location(betas.7, annot = annot.7)
cluster.vec <- Acluster(ordr.vec = dat.7.ord$betas.by.chrom[[1]], thresh.dist = 0.5, location.vec = dat.7.ord$sites.locations.by.chrom[[1]]$Coordinate_37, max.dist = 1000, type = "single")
c.mat <- cor(t(betas.7)[,1:15], method = "spearman")
plot.cor.clusts(cor.mat = c.mat, clusts  = cluster.vec[1:15])

}
