\name{print.site.analysis}
\alias{print.site.analysis}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
INTERNAL
}
\description{
Prints individual sites analysis results into a latex table. The function is used by summarize.top.clusters.
}
\usage{
print.site.analysis(res.mat, file.name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{res.mat}{
A data frame summarizing the results of the individual site analysis
}
  \item{file.name}{
File name to print the latex table to
}
}

\note{
The function returns nothing, it just prints a data frame nicely to a latex table.
}
