% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.annotation.R
\name{load.annotation}
\alias{load.annotation}
\title{Title Load annotation table}
\usage{
load.annotation(annotation.file.name)
}
\arguments{
\item{annotation.file.name}{The name of the .csv Illumina annotation file.}
}
\value{
A data.table object with the annotation.
}
\description{
The package is adjusted to Illumina annotation .csv file.
}
\examples{

annotation.file.name <- "illumina_450_manifest_v.1.2.csv"
annot <- load.annotation(annotation.file.name)
annot[1:5]

}
